+++ dse_do_utils/deployeddomodel.py
@@ -1,20 +1,23 @@
 # Copyright IBM All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
+from typing import Optional
 
 import pandas as pd
 # import sys
-from watson_machine_learning_client import WatsonMachineLearningAPIClient
+# from watson_machine_learning_client import WatsonMachineLearningAPIClient  # Deprecated
+from dse_do_utils.datamanager import Inputs
+from ibm_watson_machine_learning import APIClient  # New client
 
 
 class DeployedDOModel(object):
     """
-    New DeployedDOModel for CPD2.5
+    New DeployedDOModel for CPD3.5 based on ibm_watson_machine_learning.APIClient
 
     Major steps:
 
         1. Create an instance of a DeployedDOModel, configuring parameters
 
-    Internally, the processes uses the WatsonMachineLearningAPIClient to communicate with the deployed model:
+    Internally, the processes uses the APIClient (former WatsonMachineLearningAPIClient) to communicate with the deployed model:
 
         2. Start the solve job.
         3. Monitor the running solve job. Runs in a loop, updates the current state.
@@ -34,7 +37,7 @@ class DeployedDOModel(object):
     Usage::
 
         # Simplest, using all default options:
-        mdl = DeployedDOModel(wml_credentials, space_name, model_name)
+        mdl = DeployedDOModel(wml_credentials, space_name, deployed_model_name)
         job_details = mdl.solve(inputs)
         print("Solve status: {}".format(mdl.solve_status))
         print("Objective: {}".format(mdl.objective))
@@ -45,30 +48,42 @@ class DeployedDOModel(object):
 
         1. Debug mode
         2. Get the cplex log file?
-        3. Add kill and stop methods
-        4. Configure polling interval
+        3. Done - Add kill and stop methods
+        4. Done - Configure polling interval
 
 
     """
 
-    def __init__(self, wml_credentials, space_name=None, model_name=None, deployment_uid=None):
+    def __init__(self, wml_credentials,
+                 space_name: Optional[str]= None, deployed_model_name: Optional[str]= None, deployment_id: Optional[str]=None,
+                 default_max_oaas_time_limit_sec: Optional[int]= None, default_max_run_time_sec: Optional[int]= 600, monitor_loop_delay_sec: int = 5):
         """Initialize the interface object.
-        If the deployment_uuid is specified (WS Cloud), the space_name and model_name are optional.
+        If the deployment_uuid is specified (WS Cloud), the space_name and model_name are optional.  TODO: test on IBM Cloud
         If no deployment_uuid (CPD), specify both the model and space name.
-        Will find UUID based on space and model id.
+        Will find UUID based on space and deployed model id.
+        In CPDv3.5, always define the space_name, in combination with either the model_name, or the deployment_id
+
+        Time limits:
+        - Both are optional: if no value, no time-limit is imposed
+        - These are default values. Can be overridden in solve method
 
         Args:
-            model_name (str): name of deployed model (CPD)
+            deployed_model_name (str): name of deployed model (CPD)
             space_name (str): name of deployment space (CPD)
-            deployment_uuid (str): Deployment UUID (WS Cloud)
+            deployment_id (str): Deployment UUID (WS Cloud)
+            default_max_oaas_time_limit_sec (int): default oaas.timeLimit in seconds.
+            default_max_run_time_sec (int): default maximum run time in seconds. Includes the queueing time.
+            monitor_loop_delay_sec (int): delay in seconds in monitoring/polling loop
         """
 
         # Inputs
         self.wml_credentials = wml_credentials
         self.space_name = space_name
-        self.model_name = model_name
-        self.deployment_uid = deployment_uid
-        self.max_run_time = 60  # seconds: Job will be deleted
+        self.model_name = deployed_model_name
+        self.deployment_id = deployment_id
+        self.default_max_oaas_time_limit_sec = default_max_oaas_time_limit_sec  # In seconds! None implies no time timit. Note the original oaas.timeLimit is in milli-seconds!
+        self.default_max_run_time_sec = default_max_run_time_sec  #60  # In seconds: Job will be deleted. None implies no time timit.
+        self.monitor_loop_delay_sec = monitor_loop_delay_sec  # In seconds
         # self.time_limit = 600  # in milliseconds. timeLimit for DO model cancel
         #         self.inputs = inputs
         #         self.debug = debug
@@ -84,20 +99,28 @@ class DeployedDOModel(object):
         self.job_details = None
 
         # Setup and connection to deployed model
-        self.client = WatsonMachineLearningAPIClient(wml_credentials)
+        self.client = APIClient(wml_credentials)
 
-        if self.deployment_uid is None:
-            space_id = [x['metadata']['id'] for x in self.client.spaces.get_details()['resources'] if
-                        x['entity']['name'] == space_name][0]
-            self.client.set.default_space(space_id)
-            self.deployment_uid = [x['metadata']['guid'] for x in self.client.deployments.get_details()['resources'] if
-                                   x['entity']['name'] == model_name][0]
+        # space_id = [x['metadata']['id'] for x in self.client.spaces.get_details()['resources'] if
+        #             x['entity']['name'] == space_name][0]
+        space_id = self.get_space_id(space_name)
+        self.client.set.default_space(space_id)  # Also required when using deployment_id
 
-    def solve(self, inputs, time_limit=1000, max_run_time=60):
+        if self.deployment_id is None:
+            #             space_id = [x['metadata']['id'] for x in self.client.spaces.get_details()['resources'] if
+            #                         x['entity']['name'] == space_name][0]
+            #             self.client.set.default_space(space_id)
+            # self.deployment_uid = [x['metadata']['guid'] for x in self.client.deployments.get_details()['resources'] if
+            #                        x['entity']['name'] == model_name][0]
+            self.deployment_id = self.get_deployment_id(deployed_model_name)
+
+    def solve(self, inputs: Inputs, max_oaas_time_limit_sec: int = None, max_run_time_sec: int = None):
         """Master routine. Initializes the job, starts the execution, monitors the results, post-processes the solution and cleans-up after.
 
         Args:
             inputs (dict of DataFrames): input tables
+            max_oaas_time_limit_sec (int): will override the default from the constructor
+            max_run_time_sec (int): will override the default from the constructor
 
         Calls the following methods (in order)::
 
@@ -109,39 +132,47 @@ class DeployedDOModel(object):
             self.cleanup()
 
         """
+        if max_run_time_sec is None:
+            max_run_time_sec = self.default_max_run_time_sec
+        if max_oaas_time_limit_sec is None:
+            max_oaas_time_limit_sec = self.default_max_oaas_time_limit_sec
         from time import sleep
-        job_details = self.execute_model(inputs, time_limit)
+        job_details = self.execute_model(inputs, max_oaas_time_limit_sec)
         sleep(0.5)  # Give a little time for the job to start
-        job_details = self.monitor_execution(job_details, max_run_time)
+        job_details = self.monitor_execution(job_details, max_run_time_sec)
         self.extract_solution(job_details)
         return job_details
 
     #                 self.cleanup()
 
-    def get_solve_payload(self, inputs, time_limit):
+    def get_solve_payload(self, inputs: Inputs, max_oaas_time_limit_sec: Optional[int] = None):
         input_data = [{"id": f"{table_name}.csv", "values": df} for table_name, df in inputs.items()]
         output_data = [{"id": ".*\.csv"}]
+        solve_parameters = {"oaas.logTailEnabled": "true",
+                            "oaas.logLimit": 10000,
+                            "oaas.logAttachmentName": 'log.txt'}
+        if max_oaas_time_limit_sec is not None:
+            solve_parameters['oaas.timeLimit'] = max_oaas_time_limit_sec * 1000,  # oaas.timeLimit needs to be specified in milli-seconds
         solve_payload = {self.client.deployments.DecisionOptimizationMetaNames.INPUT_DATA: input_data,
                          self.client.deployments.DecisionOptimizationMetaNames.OUTPUT_DATA: output_data,
-                         self.client.deployments.DecisionOptimizationMetaNames.SOLVE_PARAMETERS:
-                             {"oaas.timeLimit": time_limit,
-                              "oaas.logTailEnabled": "true",
-                              "oaas.logLimit": 10000,
-                              "oaas.logAttachmentName": 'log.txt'}}
-        print(f"Time-limit = {time_limit}")
+                         self.client.deployments.DecisionOptimizationMetaNames.SOLVE_PARAMETERS: solve_parameters
+                         #                              {"oaas.timeLimit": max_oaas_time_limit_sec * 1000,  # oaas.timeLimit needs to be specified in milli-seconds
+                         #                               "oaas.logTailEnabled": "true",
+                         #                               "oaas.logLimit": 10000,
+                         #                               "oaas.logAttachmentName": 'log.txt'}
+                         }
+        print(f"max_oaas_time_limit_sec = {max_oaas_time_limit_sec}")
         return solve_payload
 
-    def execute_model(self, inputs, time_limit):
-        solve_payload = self.get_solve_payload(inputs, time_limit)
-        job_details = self.client.deployments.create_job(self.deployment_uid, solve_payload)
+    def execute_model(self, inputs: Inputs, max_oaas_time_limit_sec: Optional[int]):
+        solve_payload = self.get_solve_payload(inputs, max_oaas_time_limit_sec)
+        job_details = self.client.deployments.create_job(self.deployment_id, solve_payload)
         return job_details
 
-    def monitor_execution(self, job_details, max_run_time):
+    def monitor_execution(self, job_details, max_run_time_sec: Optional[int] = None):
         """Monitor the model execution by periodically calling the API to get the current execution status.
         Result stored in self.execution_status_json and self.execution_status.
 
-        TODO: Time-out?
-        TODO: control the loop delay?
         """
         import time
         # from time import sleep
@@ -152,18 +183,19 @@ class DeployedDOModel(object):
         job_status = DeployedDOModel.get_job_status(job_details)
         while job_status not in ['completed', 'failed', 'canceled']:
             print(f"{job_status}.... run-time={elapsed_time:.1f}")
-            time.sleep(5)
+            time.sleep(self.monitor_loop_delay_sec)
             job_details = self.client.deployments.get_job_details(job_uid)
             job_status = DeployedDOModel.get_job_status(job_details)
             elapsed_time = time.time() - start_time
-            if elapsed_time > max_run_time:
+            if max_run_time_sec is not None and elapsed_time > max_run_time_sec:
                 self.client.deployments.delete_job(job_uid, hard_delete=True)
-                print(f"Job deleted due to run-time exceeding maximum limit of {max_run_time} seconds")
+                print(f"Job deleted due to run-time exceeding maximum limit of {max_run_time_sec} seconds")
                 self.solve_status = 'JOB DELETED'
                 break
 
         self.run_time = elapsed_time
-        print(job_status)
+        #         print(job_status)
+        print(f"End monitor_execution with job_status = {job_status}, run-time={elapsed_time:.1f}")
         return job_details
 
     @staticmethod
@@ -179,7 +211,6 @@ class DeployedDOModel(object):
             self.objective = self.get_solve_details_objective(job_details)
             self.outputs = self.get_outputs(job_details)
             self.solve_details = self.get_solve_details(job_details)
-
 
     @staticmethod
     def get_outputs(job_details):
@@ -205,7 +236,14 @@ class DeployedDOModel(object):
         """After job has completed"""
         return job_details['entity']['decision_optimization']['solve_state']['details']['PROGRESS_BEST_OBJECTIVE']
 
+    def get_space_id(self, space_name: str):
+        """Find space_id from space_name."""
+        space_id = [x['metadata']['id'] for x in self.client.spaces.get_details()['resources'] if
+                    x['entity']['name'] == space_name][0]
+        return space_id
 
-
-
-
+    def get_deployment_id(self, model_name: str):
+        """Find deployment_id from model_name."""
+        deployment_id = [x['metadata']['id'] for x in self.client.deployments.get_details()['resources'] if
+                         x['entity']['name'] == model_name][0]
+        return deployment_id